requirements.txt
yfinance
pandas
numpy
sklearn
tensorflow
flask
python-telegram-bot
matplotlib
